backend/requirements.txt
# AI-First Marketing Automation - Backend Dependencies

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Data Validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# OpenAI
openai==1.10.0

# Image Processing
Pillow>=10.3.0

# HTTP Requests
requests==2.31.0
httpx==0.26.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Social Media APIs
linkedin-api==2.2.0
facebook-sdk==3.1.0

# Testing (optional)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development (optional)
black==24.1.1
flake8==7.0.0
mypy==1.8.0

